+++ CHANGELOG.md
@@ -14,6 +14,7 @@ BHT-EMR-API or eMastercard2Nart gets updated then the tag will get bumped up to 
 ### Changed
 
 - Bumped up BHT-EMR-API to v4.10.34
+- Added autoloading of metadata after loading backup
 
 ## [v4.0.17-0] 2021-04-14